% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.raref.r
\name{beta.raref}
\alias{beta.raref}
\title{Rarefaction of true beta diversity (Whittaker's multiplicative beta)}
\usage{
beta.raref(comm, sites, conditioned = TRUE, gamma = "jack1")
}
\arguments{
\item{comm}{Community matrix (or data frame) for which rarified beta diversity is calculated.}

\item{sites}{Number of sites for which rarefied true beta should be calculated.}

\item{conditioned}{Logical; should the result be conditioned on the provided dataset (\code{comm} matrix) or not? Default = \code{TRUE}.}

\item{gamma}{Estimator of gamma diversity used to estimate number of species in the community (in case that \code{conditioned = FALSE}). \code{gamma = c('chao', 'jack1', 'jack2', 'boot')}.}
}
\value{
The function returns list with the following items:
\itemize{
\item \code{sites}: number of sites for which rarefaction is calculated;
\item \code{gamma}: estimated number of species (gamma diversity) for given number of \code{sites};
\item \code{gamma.sd}: standard deviation of \code{gamma} estimate;
\item \code{alpha}: mean number of species per sample (alpha diversity);
\item \code{beta}: true beta diversity, calculated as \code{gamma / alpha};
\item \code{beta.sd}: standard deviation of true beta diversity, calculated as \code{gamma.sd / alpha} (perhaps wrongand better not to use, needs proof!);
}
}
\description{
Calculates true beta diversity rarefied to given number of samples.
}
\examples{
sc <- sample.comm (simul.comm (totS = 100), Np= 100)
beta.raref (comm = sc$a.mat, sites = 10)

# True beta diversity calculated on pair of samples (minus one) 
# is very close to mean pairwise Sorensen:
beta.raref (comm = sc$a.mat, sites = 2)$beta-1
library (vegan)
mean (vegdist (decostand (sc$a.mat, 'pa')))
}
\author{
David Zeleny (zeleny.david@gmail.com). Heavily builds on the code from function \code{specaccum} in package \code{vegan} written by Roeland Kindt and Jari Oksanen to construct sample-based rarefaction curves.
}
