% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_node_centrality_scores.R
\name{Borda}
\alias{Borda}
\title{Borda Based Rank Aggregation}
\usage{
Borda(input)
}
\arguments{
\item{input}{a list containing individual ranked lists.}
}
\value{
a list with two components:\cr
-\strong{TopK} a matrix with 4 columns corresponding to rankings by each of the 4 aggregation functions;\cr
-\strong{Scores} a matrix with 4 columns corresponding to the Borda scores from each of the 4 aggregation functions.
}
\description{
Computes Borda scores and rank based on four different aggregation functions.
}
\details{
Computes Borda scores and rank based on four different aggregation functions. The four aggregation functions
are mean, median, geometric mean and L2 norm.
}
\keyword{internal}
