% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_tissue_specific_efficacy_score.R
\name{build.tissue.specific.networks}
\alias{build.tissue.specific.networks}
\title{Compile tissue-specific networks and, within these, set of genes that are relevant for the selected targets.}
\usage{
build.tissue.specific.networks(
  tissue_scores,
  disease_genes,
  ppi_network,
  directed_network = F,
  tissue_expr_data,
  top_targets = NULL,
  rwr_restart = 0.75,
  rwr_norm = "quantile",
  rwr_cutoff = 0.001,
  verbose = FALSE
)
}
\arguments{
\item{tissue_scores}{a data.frame as the one compiled by \code{get.tissue.specific.scores}}

\item{disease_genes}{character vector containing the IDs of the genes related to a particular disease.
Gene IDs are expected to match with those provided in \code{ppi_network} and \code{tissue_expr_data}.}

\item{ppi_network}{a matrix or a data frame with at least two columns
reporting the ppi connections (or edges). Each line corresponds to a direct interaction.
Columns give the gene IDs of the two interacting proteins.}

\item{directed_network}{logical indicating whether the PPI is directed.}

\item{tissue_expr_data}{a numeric matrix or data frame indicating expression significances
in the form of Z-scores. Columns are tissues and rows are genes; colnames and rownames must be provided.
Gene IDs are expected to match with those provided in \code{ppi_network}.}

\item{top_targets}{character vector indicating a list of ENTREZ id to be used for the slection of the shortest paths.}

\item{rwr_restart}{the restart probability used for RWR. See \code{dnet::dRWR} for more details.}

\item{rwr_norm}{the way to normalise the adjacency matrix of the input graph. See \code{dnet::dRWR} for more details.}

\item{rwr_cutoff}{the cuoff value to select the most visited genes.}

\item{verbose}{logical indicating whether the messages will be displayed or not in the screen.}
}
\value{
a list of four objects: \cr
       - \strong{tsn}: a list of tissue-specific networks;\cr
       - \strong{shp}: a list of gene sets, each gene set indicates the genes connecting a target to all disease genes;\cr
       - \strong{tsn}: a list of gene sets, each gene set represents the set of genese that are closely related to the set of targets;\cr
       - \strong{universe}: the total number of genes in the tissue-specific networks.
}
\description{
Given a set of gene targets, it builds the corresponding tissue-specific networks, the set of genes linking the targets to
disease genes, a set of genes tightly connected to all targets.
}
\details{
The top targets are used to re-build the shortest paths with the disease-relevant genes in tissue-specific networks.
The shortest paths linkining a top target to disease genes are merged and the resulting set of nodes/genes are giving in output.
Moreover, random walk with restart is utilized to identify a set of genes that is tightly connected to the targets.
}
