% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{centrality_score}
\alias{centrality_score}
\title{Node centrality scores for the tissue-specific efficacy scores}
\format{A list containing node centrality scores.}
\usage{
data(centrality_score)
}
\description{
Three node centrality scores are measured (node degree, clustering coefficient and betweenness)
for each tissue-specific network. The networks are built by combining tissue-specific gene expression profiles
(which are obtained from GTEx; \url{https://gtexportal.org/}) and the human protein-protein interaction
(which is obtained from StringDB; \url{https://string-db.org/}). The Borda method is used to merge the three ndoe centrality scores.
These scores are finally discretized by replacing the rank values with discrete levels of node centrality: 0.25, 0.5, 0.75 and 1.
}
\details{
It is a list of 4 items:
   - \code{scores}: three centrality scores compiled for each <gene, tissue> pair (list of matrices with genes on the rows and scores on the columns);
   - \code{ranks}: node rankings (list of matrices with genes on the rows and ranks on the columns);
   - \code{borda}: the borda count merging the node rankings (a matrix with genes on the rows and the borda-based rank on the columns/tissues);
   - \code{borda}: a list of discretized borda counts for each tissue.
}
\examples{
data(centrality_score)
names(centrality_score)
head(centrality_score$scores[[1]])
table(centrality_score$borda.disc[[1]])
}
\references{
\insertRef{Failli2019}{ThETA}
}
\keyword{datasets}
