% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_disease_relevant_tissues.R
\name{dis.rel.tissues}
\alias{dis.rel.tissues}
\alias{list.dis.rel.tissues}
\title{Significance Of Disease Modules In Tissue-Specific Interactomes}
\usage{
dis.rel.tissues(disease_genes, ppi_network, weighted = FALSE,
  tissue_expr_data, thr = 1, top = 20, rand = 100, verbose = FALSE)

list.dis.rel.tissues(disease_gene_list, ppi_network, weighted = FALSE,
  tissue_expr_data, thr = 1, top = 20, rand = 100, parallel = NULL)
}
\arguments{
\item{disease_genes}{character vector containing the IDs of the genes related to a particular disease.
Gene IDs are expected to match with those provided in \code{ppi_network} and \code{tissue_expr_data}.}

\item{ppi_network}{a matrix or a data frame with at least two columns
reporting the ppi connections (or edges). Each line corresponds to a direct interaction.
Columns give the gene IDs of the two interacting proteins.}

\item{weighted}{logical indicating whether the \code{ppi_network} contain edge weights? If \code{TRUE}, then the third column
is used to weight edges.}

\item{tissue_expr_data}{a numeric matrix or data frame indicating expression significances
in the form of Z-scores. Columns are tissues and rows are genes; colnames and rownames must be provided.
Gene IDs are expected to match with those provided in \code{ppi_network}.}

\item{thr}{an integer specifying the threshold for expression significance. Only genes with
expression significance >= \code{thr} will be considered expressed in a particular tissue.
Defaults to 1.}

\item{top}{an integer indicating the number of genes to be considered to extend the disease gene set.
Genes are selected based on the radial distance from the disease-relevant genes.}

\item{rand}{an integer indicating the number of random sets of genes needed to calculate the singificance of disease-tissue associations. Defaults to 100.}

\item{verbose}{logical indicating whether the messages will be displayed or not in the screen.}

\item{disease_gene_list}{a list of disease-associated genes. Each element of the list is a character vector
containing the IDs of the genes related to a particular disease.}

\item{parallel}{an integer indicating how many cores will be registered for parallel computation.
If \code{NULL}, the function will not use parallel computation.}
}
\value{
A list of five items:\cr
       - \strong{maxc}: a vector or a matrix specifying the disease maximal connected components by tissue;\cr
       - \strong{mer}: a vector or a matrix specifying the averages of \code{rand} maximal connected components by tissue;\cr
       - \strong{sdr}: a vector or a matrix specifying the standard deviations of \code{rand} maximal connected components by tissue;\cr
       - \strong{z}: a vector or a matrix specifying the significances of disease-tissue associations in the form of Z-scores;\cr
       - \strong{genes}: a vector or a matrix specifying the cardinality of disease-relevant genes expressed in each tissue.
}
\description{
Compile significance of disease modules in tissue-specific interactomes.
}
\details{
This function implements the method from \insertRef{Kitsak2016}{ThETA}.
The function tests the significance of disease modules in tissue-specific interactomes, consisting
of genes with expression significance >= \code{thr}. First, disease maximal connected
component is measured in tissue \emph{t}. Then, the average of \code{rand} maximal connected
components in \emph{t} is determined to compute the significance of real data.
As a major improvement, the radial distance between each gene and all disease-relevant genes is computed
to extend the disease gene set.
}
