% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{dis_vrnts}
\alias{dis_vrnts}
\title{Disease-associated genes retrieved from DIsGeNET.}
\format{A list of data frames.}
\usage{
data(dis_vrnts)
}
\description{
Disease-associated genes retrieved from DIsGeNET (\url{www.disgenet.org}).
}
\details{
It corresponds to a list of data frames. Each data frame contains three columns reporting entrez id,
symbol and  score of genes significantly associated with a disease. Each entry (or disease)
in the list is indexed by an EFO-ID.
}
\examples{
data(dis_vrnts)
}
\references{
\insertRef{Failli2019}{ThETA}
}
\keyword{datasets}
