% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_disease_relevant_tissues.R
\name{disease.vrnts}
\alias{disease.vrnts}
\title{Retrieving Gene-Disease Assocations From DisGeNET}
\usage{
disease.vrnts(x, id_type = "efo", min_score = 0, curated = F)
}
\arguments{
\item{x}{character indicating the disease ID (eg. efo ID).}

\item{id_type}{character indicating the type of disease ID. Possible values are:
\code{efo}, \code{mesh} and \code{omim}. Defaults to \code{efo}.}

\item{min_score}{numeric value from 0 to 1 indicating the cut-off for gene-disease scores. Defaults to 0.}

\item{curated}{a boolean value indicating if manually curated gene-disease pairs must be selected.}
}
\value{
a three columns data frame:\cr
       - \strong{entrez}: a character variable containing entrez gene ids;\cr
       - \strong{symbol}: a character variable containing gene symbols;\cr
       - \strong{score}: a numeric variable containing gene-disease scores.
}
\description{
Utility function to extract gene-disease association from DisGeNET \url{http://www.disgenet.org/}.
}
