% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{disease_tissue_zscores}
\alias{disease_tissue_zscores}
\title{Disease-associated tissue scores.}
\format{A list of numerical matrices.}
\usage{
data(disease_tissue_zscores)
}
\description{
Significance of disease modules in tissue-specific interactomes.
}
\details{
Significances were obtained by implementing the algorithm described in \insertRef{Kitsak2016}{ThETA},
which aims to assess tissue specificity of human disease modules.
It is a list of 5 matrices with EFO-id on the rows and GTEx-based tissues on the clomuns:
       - \code{maxc}: disease maximal connected components.
       - \code{mer}: averages of random maximal connected components.
       - \code{sdr}: standard deviations of random maximal connected components.
       - \code{z}: significances of disease-tissue associations in the form of Z-scores.
       - \code{genes}: cardinality of disease-relevant genes expressed in each tissue.
}
\examples{
data(disease_tissue_zscores)
names(disease_tissue_zscores)
head(disease_tissue_zscores$z)
}
\references{
\insertRef{Failli2019}{ThETA}
}
\keyword{datasets}
