% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genspe.r
\docType{package}
\name{genspe}
\alias{genspe}
\alias{genspe-package}
\title{genspe - package for calculating and testing co-occurrence based measure of species habitat specialization, introduced by Fridley et al. 2007}
\description{
genspe - package for calculating and testing co-occurrence based measure of species habitat specialization, introduced by Fridley et al. 2007
}
\author{
David Zeleny
}

