% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_node_centrality_scores.R
\name{geo.mean}
\alias{geo.mean}
\title{Calculate The Geometric Mean}
\usage{
geo.mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric, a vector of values.}

\item{na.rm}{logical, whether or not to remove NA values from the calculation.}
}
\value{
the geometric mean of \code{x}
}
\description{
Calculate the geometric mean.
}
\keyword{internal}
