% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{geo_gene_sets}
\alias{geo_gene_sets}
\title{Lists of gene sets induced by disease and gene perturbations.}
\format{A numerical matrix.}
\usage{
data(geo_gene_sets)
}
\description{
These lists were retrieved from EnrichR (\url{https://amp.pharm.mssm.edu/Enrichr/}) and formatted
in order to contain two items:
       - \code{geneIds}: a vector of gene IDs up- or down-regulated in diseases or gene perturbations
       - \code{Description}: a character specifying either the extended name of the disease or the gene modulation type.
}
\examples{
data(geo_gene_sets)
}
\references{
\insertRef{Kuleshov2016}{ThETA}
}
\keyword{datasets}
