% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{GS_CTD}
\alias{GS_CTD}
\title{Set of known drug-target-disease associations retrieved from the Comparative Toxicogenomics Database (CTD)
(http://ctdbase.org/)}
\format{A data frame indicating gene(drug-target)-disease associations.

  target.entrez      disease.id                        description
           7039     EFO:0003884             chronic kidney disease
           1490     EFO:0009549 female reproductive system disease}
\usage{
data(GS_CTD)
}
\description{
Set of known drug-target-disease associations retrieved from the Comparative Toxicogenomics Database (CTD)
(http://ctdbase.org/)
}
\examples{
data(GS_CTD)
}
\keyword{datasets}
