% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{GS_DBK}
\alias{GS_DBK}
\title{Set of known drug-target-disease associations retrieved from the DrugBank Database (DBK)
(https://www.drugbank.ca/)}
\format{A data frame indicating gene(drug-target)-disease associations.

  target.entrez      disease.id                        description
           7039     EFO:0003884             chronic kidney disease
           1490     EFO:0009549 female reproductive system disease}
\usage{
data(GS_DBK)
}
\description{
Set of known drug-target-disease associations retrieved from the DrugBank Database (DBK)
(https://www.drugbank.ca/)
}
\examples{
data(GS_DBK)
}
\keyword{datasets}
