% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{GS_TTD}
\alias{GS_TTD}
\title{Set of known drug-target-disease associations retrieved from the Therapeutic Target Database (TTD)
(http://db.idrblab.net/ttd/)}
\format{A data frame indicating gene(drug-target)-disease associations.

  target.entrez      disease.id                        description
           7039     EFO:0003884             chronic kidney disease
           1490     EFO:0009549 female reproductive system disease}
\usage{
data(GS_TTD)
}
\description{
27/02/2020
}
\examples{
data(GS_TTD)
}
\keyword{datasets}
