% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{gtexv7_zscore}
\alias{gtexv7_zscore}
\title{Tissue-specific gene expression profiles retrieved from GTEx.}
\format{A numerical matrix.}
\usage{
data(gtexv7_zscore)
}
\description{
Z-score of log(TPM+1) values.
}
\examples{
data(gtexv7_zscore)
head(gtexv7_zscore)
}
\references{
\insertRef{GTEx2013}{ThETA}
}
\keyword{datasets}
