% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_node_centrality_scores.R
\name{l2norm}
\alias{l2norm}
\title{Calculate The L2 Norm}
\usage{
l2norm(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric, a vector of values.}

\item{na.rm}{logical, whether or not to remove NA values from the calculation.}
}
\value{
The L2 norm of \code{x}
}
\description{
Calculate the l2 norm.
}
\keyword{internal}
