% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_ppv.R
\name{matching.rows}
\alias{matching.rows}
\title{A function for matching rows between two matrices or data.frames.}
\usage{
matching.rows(dat, tab, nomatch = NA)
}
\description{
A function for matching rows between two matrices or data.frames.
}
\keyword{internal}
