% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_modulation_score.R
\name{modulation.score}
\alias{modulation.score}
\title{Modulation Scores Compiled Between Gene Modulations And Disease Perturbations.}
\usage{
modulation.score(geneSets = NULL, verbose = FALSE)
}
\arguments{
\item{geneSets}{a list of four lists including up or down gene sets identified in disease or gene perturbations
(e.g., gene knockout, gene knockdown, etc.). Lists of disease and gene perturbations shall not be shuffled and
the same ordering criterion shall be applied (e.g. disease-up-gene-sets, disease-down-gene-sets, gene-up-gene-sets and gene-down-gene-sets).
Each gene set object is a list of two items:\cr
       - \code{geneIds}: a vector of gene IDs up- or down-regulated in diseases or gene perturbations;\cr
       - \code{Description}: a character specifying either the extended name of the disease or the gene modulation type.
If the list object is empty, then the following list of genes sets will be downloaded from EnrichR
(\url{https://amp.pharm.mssm.edu/Enrichr/}) by using the R (Bioconductor) package \pkg{MIGSA} \insertRef{Rodriguez2019}{ThETA}.}

\item{verbose}{logical indicating whether the messages will be displayed or not in the screen.}
}
\value{
a data frame with target-disease associations and related modulation scores. The gene modulation type and
the disease name are also reported.
}
\description{
Determine if a gene modulation can revert disease-induced gene expression patterns.
}
\details{
This function implements the core of the modulation (efficacy) score described in \insertRef{Failli2019}{ThETA}.
}
