% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_node_centrality_scores.R
\name{node.centrality}
\alias{node.centrality}
\title{Compile Node Centrality Scores For A Given PPI Network.}
\usage{
node.centrality(
  ppi_network,
  tissue_expr_data,
  agg_function = 4,
  directed_network = F,
  parallel = NULL,
  verbose = FALSE
)
}
\arguments{
\item{ppi_network}{a matrix or a data frame with at least two columns
reporting the ppi connections (or edges). Each line corresponds to a direct interaction.
Columns give the gene IDs of the two interacting proteins.
Only interactions whose destination nodes match with the IDs provided in \code{tissue_expr_data}
will be considered during the computation.}

\item{tissue_expr_data}{numeric matrix or data frame indicating expression significances
in the form of Z-scores. Columns are tissues and rows are genes; colnames and rownames must be provided.}

\item{agg_function}{integer value between 1 and 4. Provides control over Borda aggregation functions.
1 for \code{mean}, 2 for \code{median}, 3 for \code{geo.mean} and 4 for \code{l2norm}. Defaults to 4.}

\item{directed_network}{boolean value indicating the use of a direct network. Defaults to FALSE.}

\item{parallel}{an integer indicating how many cores will be registered for parallel computation.
If NULL, the function will not use parallel computation.}

\item{verbose}{logical to indicate whether the messages will be displayed or not in the screen.}
}
\value{
a list of 4 items:\cr
   - \strong{scores}: three centrality scores compiled for each <gene, tissue> pair (list of matrices with genes on the rows and scores on the columns);\cr
   - \strong{ranks}: node rankings (list of matrices with genes on the rows and ranks on the columns);\cr
   - \strong{borda}: the borda count merging the node rankings (a matrix with genes on the rows and the borda-based rank on the columns/tissues);\cr
   - \strong{borda}: a list of discretized borda counts for each tissue.
}
\description{
Determine three node centrality measures (node degree, clustering coefficient and betweenness) for a given ppi network.
}
\details{
The three scores are merged with the Borda method. See \insertRef{Failli2019}{ThETA}, for more details.
}
