% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_tissue_specific_efficacy_score.R
\name{novelty.plots}
\alias{novelty.plots}
\title{Generate PubMed Central Trend plots.}
\usage{
novelty.plots(list_genes, gene_id = "ENTREZID", orgdb = NULL,
  pubmed = c(2010, 2019), font_size = 8)
}
\arguments{
\item{list_genes}{a list of gene names.}

\item{gene_id}{type of gene-ID.}

\item{orgdb}{a character specifying the organism for the mapping between gene identifiers.}

\item{pubmed}{period of query in the unit of year.}

\item{font_size}{text size in pts.}
}
\value{
a PubMed-trend plot.
}
\description{
It uses the pmcplot function from R package 'enrichplot' to generate gene-based trend plots.
}
