% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{ppi_strdb_700}
\alias{ppi_strdb_700}
\title{Human protein-prtein interaction network retrieved from StringDB.}
\format{A numerical matrix.}
\usage{
data(ppi_strdb_700)
}
\description{
This ppi was built by considering only connections with a combined score greater or equal than 700.
}
\examples{
data(ppi_strdb_700)
head(ppi_strdb_700)
}
\references{
\insertRef{Franceschini2013}{ThETA}
}
\keyword{datasets}
