% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_ppv.R
\name{ppvpercents}
\alias{ppvpercents}
\title{Plotting function showing positive predictive values.}
\usage{
ppvpercents(
  gold_std,
  test,
  entrez.col = NULL,
  disease.col = NULL,
  score.col = NULL,
  a.disease = NULL,
  plot = TRUE
)
}
\arguments{
\item{gold_std}{a dataframe representing known drug target disease associations. 
It must contain at least two columns indicating the gene-entrez-id and disease-efo-id (EFO-ID). 
Ex:
target.entrez      disease.id                        description
        7039     EFO:0003884             chronic kidney disease
        1490     EFO:0009549 female reproductive system disease}

\item{test}{a dataframe containing efficacy estimates of disease-drug-target(gene) associations.
It must contain at least three columns indicating the gene-entrez-id, disease-efo-id (EFO-ID) and the efficacy scores. 
Ex:
target.entrez   disease.id  modscore
         5376 EFO:0000270 0.4614037
        28996 EFO:0000270 0.5404949
         7403 EFO:0000270 0.6425815}

\item{entrez.col}{a character value indicating the column for the gene-entrez-id.}

\item{disease.col}{a character value indicating the column for the disease-efo-id.}

\item{score.col}{a character value indicating the column for the efficacy estimates.}

\item{a.disease}{a character value to plot the ppv values for a given disease.}

\item{plot}{a boolean value to plot the ppv curves.}
}
\value{
a ggplot showing curves of positive predictive values (PPVs).
}
\description{
Given a set of knonw disease-drug-target(gene) associations and efficacy estimates compiled for diseas-gene associations,
it determines an visualize the positive predictive values (ppv). PPVs are computed as the percentage of true positives 
normalized against the expected ppv for a random ordering of disease-gene pairs.
}
\details{
This function was used for the comparison of different efficacy scoring methods described in \insertRef{Failli2019}{ThETA}.
}
