% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simul.comm.r
\name{simul.comm}
\alias{sample.comm}
\alias{simul.comm}
\title{Simulation of ecological data structured by single ecological gradient}
\usage{
simul.comm(totS = 300, gr.length = 5000, niche.type = "random",
  max.niche.breath = gr.length, min.niche.breath = 10,
  spec.optima = "random", prop.random.species = 0, seed = NULL,
  plotting = F, highlight.species = NULL)

sample.comm(simul.comm = NULL, Np = 300, sample.x = "random",
  no.ind = 100, k = 0.2, seed = NULL, pa = FALSE,
  based.on = "individuals", standardize.rowsums = TRUE)
}
\arguments{
\item{totS}{total number of species in simulation}

\item{gr.length}{length of the gradient in units}

\item{niche.type}{shape of species response curves ('random', 'normal', 'skewed')}

\item{max.niche.breath}{number, maximum niche breath along the gradient (default = \code{gr.length})}

\item{min.niche.breath}{number, minimum niche breath along the gradient (default = \code{10}). Cannot be more than \code{max.niche.breath}.}

\item{spec.optima}{distribution of species optima along simulated gradient. Either one of \code{'random'}, \code{'unimodal'} or \code{'skewed'}, or the vector (of the length equal to \code{totS}) with numbers (ranging between 1 and \code{gr.length}). Default = \code{'random'}.}

\item{prop.random.species}{Proportion of random species, i.e. species which are not related to the gradient. Should be real number from [0,1]. These species are generated by randomizing occurrence of given proportion of species among samples. Default = 0.}

\item{seed}{set random seed for reproducing always the same result}

\item{plotting}{should the species response curves along simulated gradient be drawned? Default = \code{FALSE}.}

\item{highlight.species}{vector of species numbers which should be highlighted by color in ploted diagram (if \code{plotting = TRUE}). Default = \code{NULL}, which means that if plot is drawned, two species will be highlighted - the most generalized and the most specialized ones.}

\item{simul.comm}{result of simul.comm function with parameters of individual species response curves}

\item{Np}{number of samples}

\item{sample.x}{positions of sampling along the gradient. Default = \code{"random"}, meaning that random locations are generated. Other options include \code{"equal"} with samples distributed in equal distances, \code{"biased"} with samples accumulated toward higher values of the gradient. Can be also vector of the same length as \code{Np} with exact positions of the samples.}

\item{no.ind}{mean number of individuals to be drawn from the species pool (if \code{based.on = 'individuals'})}

\item{k}{mean proportion of species from species pool to be drawn into local community (if \code{based.on = 'species'})}

\item{pa}{result table will be generated in presence-absence form (default \code{pa = FALSE})}

\item{based.on}{should the sampling of species from species pool be based on \code{'individuals'} or \code{'species'}? (default = \code{'individuals'})}

\item{standardize.rowsums}{Standardizes the abundances of species in samples so as the sum of species abundances in sample is 100 (standardizes to rowsums equal to 100). Applies only if \code{pa = FALSE}. Default = \code{TRUE}.}
}
\value{
The function \code{simul.comm} returns \code{list} with 10 items, describing the set of parameters used to simulate community of species response curves:
\itemize{
\item \code{totS} Total number of species in simulation.
\item \code{gr.length} Length of the gradient.
\item \code{niche.type} Shape of species response curves.
\item \code{Ao} Vector of species amplitudes for the simulated gradient (heights of species response curves, corresponding to maximum probability of species to be selected to community in species optima).
\item \code{m} Vector of species optima along the simulated gradient.
\item \code{r} Vector of generated species ranges along the simulated gradient (generated niche breaths).
\item \code{range} Vector of realised species ranges along the simulated gradient, considering the truncation of species response curves by gradient margins (these differ from \code{r} especially at the gradient margins, where the generated species niche may be wide, but since the margin cuts the species occurrences, realised species niche is narrower.)
\item \code{a, g} Vectors of shape parameters for curves (used in beta function to generate the shape of the species response curve)
\item \code{A.all} Matrix (dim = gradient length x number of species) with simulated probabilites of individual specie at individual location along the simulated gradient.
}
The function \code{sample.comm} returns \code{list} of 5 items with parameters of generated community data; the last item contains also all items returned by \code{simul.comm} function:
\itemize{
\item \code{a.mat} Matrix (sample x species) of species abundances in samples.
\item \code{p.mat} Matrix (sample x species) of species occurrence probabilities in samples.
\item \code{sample.x} Vector with positions of samples along the first and second simulated gradient, respectively (environmental variable).
\item \code{sample.comm} List of 7 items storing initial settings of arguments in \code{sample.comm} (namely arguments \code{Np, based.on, no.ind, k, seed, pa} and \code{standardize.rowsums}).
\item \code{simul.comm} List of 10 items returned by function \code{simul.comm}.
}
}
\description{
Simulation of ecological data structured by single ecological gradient
}
\author{
David Zeleny (zeleny.david@gmail.com) - based on the scripts of Jason Fridley (Fridley et al. 2007, Appendix S2) and David Zeleny (Zeleny 2009, Appendix S1)
}

