% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theta.R
\docType{package}
\name{theta}
\alias{theta}
\alias{theta-package}
\title{theta - package for calculating and testing co-occurrence based measure of species habitat specialization, introduced by Fridley et al. 2007}
\description{
theta - package for calculating and testing co-occurrence based measure of species habitat specialization, introduced by Fridley et al. 2007
}
\author{
David Zeleny
}

