% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_disease_relevant_tissues.R
\name{tissue.expr}
\alias{tissue.expr}
\title{Compiling Z-scores From Gene-Level Expression Profiles}
\usage{
tissue.expr(tissue_expr_data)
}
\arguments{
\item{tissue_expr_data}{a data matrix containing gene expression values.}
}
\value{
a numeric matrix of gene expression values in the form of Z-scores. Columns are tissues and rows are diseases.
}
\description{
Utility function to determine the Z-scores for tissue-specific gene expression profiles.
}
