% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_tissue_specific_efficacy_score.R
\name{tissue.specific.scores}
\alias{tissue.specific.scores}
\alias{list.tissue.specific.scores}
\title{Compile tissue-specific scores for a given disease.}
\usage{
tissue.specific.scores(
  disease_genes,
  ppi_network,
  directed_network = F,
  tissue_expr_data,
  dis_relevant_tissues,
  W,
  cutoff = 1.6,
  verbose = FALSE
)

list.tissue.specific.scores(
  disease_gene_list,
  ppi_network,
  directed_network = F,
  tissue_expr_data,
  dis_relevant_tissues,
  W,
  cutoff = 1.6,
  verbose = FALSE,
  parallel = NULL
)
}
\arguments{
\item{disease_genes}{character vector containing the IDs of the genes related to a particular disease.
Gene IDs are expected to match with those provided in \code{ppi_network} and \code{tissue_expr_data}.}

\item{ppi_network}{a matrix or a data frame with at least two columns
reporting the ppi connections (or edges). Each line corresponds to a direct interaction.
Columns give the gene IDs of the two interacting proteins.}

\item{directed_network}{logical indicating whether the PPI is directed.}

\item{tissue_expr_data}{a numeric matrix or data frame indicating expression significances
in the form of Z-scores. Columns are tissues and rows are genes; colnames and rownames must be provided.
Gene IDs are expected to match with those provided in \code{ppi_network}.}

\item{dis_relevant_tissues}{a named numeric vector in case of \code{tissue.specific.scores} or a
numeric matrix in case of \code{list.tissue.specific.scores} indicating the significances of
disease-tissue associations in the form of Z-scores. Vector names correspond to tissue; matrix colnames and
rownames correspond to tissues and diseases, respectively. Names must be provided.}

\item{W}{a list of discretized Borda-aggregated rankings for each tissue as the one compiled by \code{get_node_centrality}.}

\item{cutoff}{numeric value indicating the cut-off for the disease-associated tissue scores.}

\item{verbose}{logical indicating whether the messages will be displayed or not in the screen.}

\item{disease_gene_list}{a list of disease-associated genes. Each element of the list is a character vector
containing the IDs of the genes related to a particular disease.}

\item{parallel}{an integer indicating how many cores will be registered for parallel computation.}
}
\value{
a data frame or a list of data frames containing tissue specific scores.
}
\description{
Determine the tissue-specific scores for genes connected to disease-associated genes within relevant tissue-specific PPIs.
}
\details{
This function implements the tissue-specific efficacy estimates of target-disease associations described in \insertRef{Failli2019}{ThETA}.
This function use \code{weighted.shortest.path} to compile weighted shortest paths connecting a gene to specified
disease-associated genes within relevant tissue-specific PPIs. The interquartile range (IQR) is applied to remove outliers
and the final set of efficacy scores are rescaled between 1 and 0, with 1 indicating the most effective targets.
}
