% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{tss_t2d_top50}
\alias{tss_t2d_top50}
\title{Lists of top 50 genes (derived from the tissue-specific score) for T2D}
\format{A numerical matrix.}
\usage{
data(tss_t2d_top50)
}
\description{
Lists of top 50 genes (derived from the tissue-specific score) for T2D
}
\examples{
data(tss_t2d_top50)
}
\keyword{datasets}
