% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_tissue_specific_efficacy_score.R
\name{visualize.graph}
\alias{visualize.graph}
\title{Interactive visualization of tissue-specific networks.}
\usage{
visualize.graph(
  tissue_scores,
  disease_genes,
  ppi_network,
  directed_network = FALSE,
  tissue_expr_data,
  top_target = NULL,
  db = "kegg",
  verbose = FALSE
)
}
\arguments{
\item{tissue_scores}{a data.frame as the one compiled by \code{get.tissue.specific.scores}}

\item{disease_genes}{character vector containing the IDs of the genes related to a particular disease.
Gene IDs are expected to match with those provided in \code{ppi_network} and \code{tissue_expr_data}.}

\item{ppi_network}{a matrix or a data frame with at least two columns
reporting the ppi connections (or edges). Each line corresponds to a direct interaction.
Columns give the gene IDs of the two interacting proteins.}

\item{directed_network}{logical indicating whether the PPI is directed.}

\item{tissue_expr_data}{a numeric matrix or data frame indicating expression significances
in the form of Z-scores. Columns are tissues and rows are genes; colnames and rownames must be provided.
Gene IDs are expected to match with those provided in \code{ppi_network}.}

\item{top_target}{character vector indicating a list of ENTREZ id to be used for the slection of the shortest paths.}

\item{db}{character indicating the database to consider for enrichment analysis.
Possible values are: \code{kegg}, \code{BP}, \code{MF} and \code{CC}. Defaults to \code{kegg}.}

\item{verbose}{logical indicating whether the messages will be displayed or not in the screen.}
}
\description{
It uses visNetwork to build tissue-specific networks on shortest paths connecting disease genes with selected drug targets(genes).
}
