% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_tissue_specific_efficacy_score.R
\name{weighted.shortest.path}
\alias{weighted.shortest.path}
\title{Weighted Shortest Paths compiled between a Gene and Disease-associated Genes}
\usage{
weighted.shortest.path(disease_genes, ppi_network, directed_network = F,
  tissue_expr_data, dis_relevant_tissues, W, cutoff = 1.6, verbose = F)
}
\arguments{
\item{disease_genes}{character vector containing the IDs of the genes related to a particular disease.
Gene IDs are expected to match with those provided in \code{ppi_network} and \code{tissue_expr_data}.}

\item{ppi_network}{a matrix or a data frame with at least two columns
reporting the ppi connections (or edges). Each line corresponds to a direct interaction.
Columns give the gene IDs of the two interacting proteins.}

\item{directed_network}{logical indicating whether the PPI is directed.}

\item{tissue_expr_data}{a numeric matrix or data frame indicating expression significances
in the form of Z-scores. Columns are tissues and rows are genes; colnames and rownames must be provided.
Gene IDs are expected to match with those provided in \code{ppi_network}.}

\item{dis_relevant_tissues}{a named numeric vector indicating the significances of disease-tissue associations in the form of Z-scores.
Names correspond to tissues.}

\item{W}{a list of discretized Borda-aggregated rankings for each tissue as the one compiled by \code{get_node_centrality}.}

\item{cutoff}{numeric value indicating the cut-off for the disease-associated tissue scores.}

\item{verbose}{logical indicating whether the messages will be displayed or not in the screen.}
}
\value{
A list of two score objects:\cr
       - \strong{shortest_paths}: a shortest path score for each pair <gene, tissue>;\cr
       - \strong{shortest_paths_avg}: average of shortest path scores compiled from all relavant tissues.
}
\description{
Determine weighted shortest paths connecting a gene to specified disease-associated genes within relevant tissue-specific PPIs.
}
\details{
This function implements the core of the tissue-specific score described in \insertRef{Failli2019}{ThETA}.
}
