% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.comm.r
\name{simul.comm}
\alias{simul.comm}
\alias{sample.comm}
\alias{plot.simcom}
\alias{plot.samcom}
\title{Simulation of ecological data structured by single ecological gradient}
\usage{
simul.comm(totS = 300, gr.length = 5000, niche.type = "random",
  max.niche.breath = gr.length, min.niche.breath = 10, amplitude = NULL,
  spec.optima = "random", prop.random.species = 0, seed = NULL,
  plotting = F, highlight.species = NULL)

sample.comm(simul.comm = NULL, Np = 300, sample.x = "random",
  no.ind = 100, cv.no.ind = 0.01, k = 0.2, seed = NULL, pa = FALSE,
  based.on = "individuals", standardize.rowsums = TRUE)

\method{plot}{simcom}(simcom, xlab = "Gradient", ylab = "Abundance", ...)

\method{plot}{samcom}(samcom, xlab = "Gradient", ylab = "Abundance", ...)
}
\arguments{
\item{totS}{Total number of species in simulation.}

\item{gr.length}{Length of the gradient in units.}

\item{niche.type}{Shape of species response curves ('random', 'normal', 'skewed').}

\item{max.niche.breath}{Maximum niche breath along the gradient (default = \code{gr.length})}

\item{min.niche.breath}{Minimum niche breath along the gradient (default = \code{10}). Cannot be more than \code{max.niche.breath}.}

\item{amplitude}{Amplitude of species response curve (height in its optimum). Default is to generate amplitude as random values from lognormal distribution (with meanlog = 2 and sdlog = 1). Can be vector of numeric values; if the vector is shorter than \code{totS}, the values are recycled. If all amplitudes should be equal, supply vector of the length one.}

\item{spec.optima}{Distribution of species optima along simulated gradient. Either one of \code{'random'}, \code{'unimodal'} or \code{'skewed'}, or a vector (of the length equal to \code{totS}) with numbers (ranging between 1 and \code{gr.length}). Default = \code{'random'}.}

\item{prop.random.species}{Proportion of random species, i.e. species which are not related to the gradient. Should be real number from [0,1]. These species are generated by randomizing occurrence of given proportion of species among samples. Default = 0.}

\item{seed}{Set random seed for reproducing always the same result.}

\item{plotting}{Should the species response curves along simulated gradient be drawned? Default = \code{FALSE}.}

\item{highlight.species}{Vector of species numbers which should be highlighted by color in ploted diagram (if \code{plotting = TRUE}). Default = \code{NULL}, which means that if plot is drawned, two species will be highlighted - the most generalized and the most specialized one.}

\item{simul.comm}{Object created by \code{simul.comm} function with parameters of individual species response curves.}

\item{Np}{Number of samples to be generated.}

\item{sample.x}{Positions of sampling locations along the gradient. Default = \code{"random"}, meaning that random locations are generated. Other options include \code{"equal"} with samples distributed in equal distances, \code{"biased"} with samples accumulated toward higher values of the gradient. Can be also vector of the same length as \code{Np} with exact positions of the samples.}

\item{no.ind}{Mean number of individuals to be sampled from the species pool (if \code{based.on = 'individuals'}).}

\item{cv.no.ind}{Coefficient of variation of number of indiduals sampled from the species pool. Increasing it allows to create set of samples with various levels of undersampling.}

\item{k}{Mean proportion of species from species pool to be drawn into local community (if \code{based.on = 'species'}). If only one number is given, in each sample given proportion of species will be sampled from species pool. If vector of two numbers (range of \code{k} valus) is given, random \code{k} in given range will be generated (with uniform distribution) for each sample; this allows to vary the degree of undersampling among generated samples.}

\item{pa}{Result table will be generated in presence-absence form (default \code{pa = FALSE}).}

\item{based.on}{Should the sampling of species from species pool be based on \code{'individuals'} or \code{'species'}? (default = \code{'individuals'})}

\item{standardize.rowsums}{Standardizes the abundances of species in samples so as the sum of species abundances in sample is 100 (standardizes to rowsums equal to 100). Applies only if \code{pa = FALSE}. Default = \code{TRUE}.}

\item{simcom}{Object returned by the function \code{simul.comm}.}

\item{xlab}{Label of the x-axis. Default to "Gradient".}

\item{ylab}{Label of the y-axis. Default to "Abundance".}

\item{...}{Other parameters passed into \code{plot}.}

\item{samcom}{Object returned by the function \code{sample.comm}.}
}
\value{
The function \code{simul.comm} returns \code{list} with 10 items, describing the set of parameters used to simulate community of species response curves:
\itemize{
\item \code{totS} Total number of species in simulation.
\item \code{gr.length} Length of the gradient.
\item \code{niche.type} Shape of species response curves.
\item \code{Ao} Vector of species amplitudes for the simulated gradient (heights of species response curves, corresponding to maximum probability of species to be selected to community in species optima).
\item \code{m} Vector of species optima along the simulated gradient.
\item \code{r} Vector of generated species ranges along the simulated gradient (generated niche breaths).
\item \code{range} Vector of realised species ranges along the simulated gradient, considering the truncation of species response curves by gradient margins (these differ from \code{r} especially at the gradient margins, where the generated species niche may be wide, but since the margin cuts the species occurrences, realised species niche is narrower.)
\item \code{a, g} Vectors of shape parameters for curves (used in beta function to generate the shape of the species response curve)
\item \code{A.all} Matrix (dim = gradient length x number of species) with simulated probabilites of individual specie at individual location along the simulated gradient. 
}
The function \code{sample.comm} returns \code{list} of 5 items with parameters of generated community data; the last item contains also all items returned by \code{simul.comm} function:
\itemize{
\item \code{a.mat} Matrix (sample x species) of species abundances in samples.
\item \code{p.mat} Matrix (sample x species) of species occurrence probabilities in samples.
\item \code{sample.x} Vector with positions of samples along the simulated gradient (environmental variable).
\item \code{sample.comm} List of 7 items storing initial settings of arguments in \code{sample.comm} (namely arguments \code{Np, based.on, no.ind, k, seed, pa} and \code{standardize.rowsums}).
\item \code{simul.comm} List of 10 items returned by function \code{simul.comm}.
}
}
\description{
Simulation of ecological data structured by single ecological gradient
}
\author{
David Zeleny (zeleny.david@gmail.com) - based on the scripts of Jason Fridley (Fridley et al. 2007, Appendix S2) and David Zeleny (Zeleny 2009, Appendix S1)
}
